Split emscripten shared header into common.h/forces.h/swe.h
Phase C prep for #354: the previous split (forces.cpp/swe.cpp/bindings.cpp)
used one shared watershed_native.h for all declarations. Break it into an
Embind-free common.h (constants, Vec3, clampf) plus forces.h and swe.h so
each translation unit only pulls in what it needs, ahead of the hull
sampling / multi-grid SWE work landing in these files.

Also tightens the Embind surface in bindings.cpp:
- register value_object<Vec3>/<WaterForceResult> before any function() that
  returns or accepts them (previously registered after, in violation of the
  documented Embind ordering rule)
- add static_assert(!is_polymorphic_v<T>) guards for both bound value types
- bump getVersion() 3 -> 4 (no signature changes) and document it in WASM.md,
  SYSTEMS.md, and the WASM water-forces docs

CMakeLists.txt / build.sh / build_colab.sh needed no changes — the
compile/link flag split and thread-free default were already correct.
